+++ addons/yat/src/commands/builtin/Set.cs
@@ -1,4 +1,5 @@
 using YAT.Helpers;
+using YAT.Overlay.Components;
 
 namespace YAT.Commands
 {